Update README.md
This is a Hotel Management System  project, in which we will have the information of the items or rooms that we have, what food items are ordered by the costumers, and how many rooms are allotted to them. Also, this project tell us about the sales and the remaining food stock we have in the hotel, and the total collection for the day.